refactor: minor adjusts on scss format and components path
