+++ src/components/RegisterForm/components/Field/index.scss
@@ -2,8 +2,6 @@
 @use '~styles/fonts.scss' as fonts;
 
 .field {
-  flex-grow: 1;
-
   &__label {
     display: block;
