+++ src/components/RegisterForm/index.scss
@@ -3,8 +3,6 @@
 @use '~styles/breakpoints.scss' as breakpoints;
 
 .register-form {
-  width: 100%;
-
   display: flex;
   flex-direction: column;
 
@@ -74,11 +72,11 @@
 
       transition: all 0.4s ease-in-out;
     }
+  }
 
-    &-img {
-      width: 2.4rem;
-      height: 2.4rem;
-    }
+  &__social-img {
+    width: 2.4rem;
+    height: 2.4rem;
   }
 
   &__divisor {
@@ -111,18 +109,18 @@
 
   &__form {
     width: 100%;
+  }
 
-    &-fields {
-      display: grid;
-      grid-template-columns: 1fr;
-      gap: 3rem;
+  &__form-fields {
+    display: grid;
+    grid-template-columns: 1fr;
+    gap: 3rem;
 
-      margin-bottom: 3rem;
-    }
+    margin-bottom: 3rem;
+  }
 
-    &-button {
-      max-width: 45rem;
-    }
+  &__form-button {
+    max-width: 45rem;
   }
 }
 
@@ -134,14 +132,12 @@
       font-size: fonts.$xl2;
     }
 
-    &__form {
-      &-fields {
-        display: grid;
-        grid-template-columns: repeat(2, 1fr);
-        gap: 3rem;
+    &__form-fields {
+      display: grid;
+      grid-template-columns: repeat(2, 1fr);
+      gap: 3rem;
 
-        margin-bottom: 6rem;
-      }
+      margin-bottom: 6rem;
     }
   }
 }
